requirements.txt
openai
ollama
gradio
python-dotenv
pandas